+++ cli/src/commands/interdiff.rs
@@ -36,10 +36,10 @@ use crate::ui::Ui;
 #[command(mut_arg("ignore_space_change", |a| a.short('b')))]
 pub(crate) struct InterdiffArgs {
     /// Show changes from this revision
-    #[arg(long, add = ArgValueCandidates::new(complete::all_revisions))]
+    #[arg(long, short, add = ArgValueCandidates::new(complete::all_revisions))]
     from: Option<RevisionArg>,
     /// Show changes to this revision
-    #[arg(long, add = ArgValueCandidates::new(complete::all_revisions))]
+    #[arg(long, short, add = ArgValueCandidates::new(complete::all_revisions))]
     to: Option<RevisionArg>,
     /// Restrict the diff to these paths
     #[arg(value_hint = clap::ValueHint::AnyPath)]