+++ cli/tests/test_global_opts.rs
@@ -642,15 +642,15 @@ fn test_help() {
     let test_env = TestEnvironment::default();
 
     let stdout = test_env.jj_cmd_success(test_env.env_root(), &["diffedit", "-h"]);
-    insta::assert_snapshot!(stdout, @r#"
+    insta::assert_snapshot!(stdout, @r"
     Touch up the content changes in a revision with a diff editor
 
     Usage: jj diffedit [OPTIONS]
 
     Options:
       -r, --revision <REVISION>  The revision to touch up
-          --from <FROM>          Show changes from this revision
-          --to <TO>              Edit changes in this revision
+      -f, --from <FROM>          Show changes from this revision
+      -t, --to <TO>              Edit changes in this revision
           --tool <NAME>          Specify diff editor to be used
           --restore-descendants  Preserve the content (not the diff) when rebasing descendants
       -h, --help                 Print help (see more with '--help')
@@ -665,7 +665,7 @@ fn test_help() {
           --quiet                        Silence non-primary command output
           --no-pager                     Disable the pager
           --config-toml <TOML>           Additional configuration options (can be repeated)
-    "#);
+    ");
 }
 
 #[test]